Introduce content lists.
Rather than holding page content in the form of an array of
span_t's and an array of image_t's, introduce a content_t type.

content_t's hold both a type and next/prev pointers, so content
can be held as a doubly linked list of different content types.

Both span_t's and image_t's are "derived" from content_t's.

Each content_t list starts (and ends) with a content_root node.
This allows us to find both the beginning and the end of the
list instantly.

Any given content_t node can be instantly unlinked from its list
without needing a root pointer (or needing to search for one).

No functional changes to extract so far; this commit just moves
subpages to use content_t lists for the main content, and for lines.