Update handling of ctm/trm, and line/span matching code.
Previously, we were fed both ctm and trm, and the code assumed that
the trm allowed us to get information about the font size etc. This
is a dangerous assumption, as any combination of ctm and trm can be
used such that when composed together they get the same result.

Accordingly, we now just take a single matrix for a span (the
composed 4x4 a/b/c/d matrix for the combined ctm/trm). Each char
then comes with an x/y value, which is the final x/y value to make
the full 2x3 ctm matrix for that char.

As we update the code to cope with this change in represenation,
we also update the span and paragraph matching code. Firstly, we
no longer rely on ctms matching exactly for lines to be 'compatible',
but rather will accept any two ctms as being compatible if they
both transform the unit direction vector into two colinear vectors.

We can therefore join spans/lines dependent on whether the
perpendicular distance between the two transformed baseline vectors
is suitably low, and whether the colinear distance is not too large.