C-style revamp.
I am incapable/unwilling of mainting the existing C-style (not
least because Visual C hates it, and keeps showing documentation
for the wrong function).

Therefore, rather than drip feeding in changes to whitespace etc
over multiple commits as I change code, blitz the whole lot at once.

There is bound to be stuff that I've missed, but this is the majority
of it at least.