Avoid breaking lines repeatedly if they start with space.
The current code, faced with:

  "bibble"
  "   foo"

would result in:

 <span>bibble</span>
 <span> </span><span> </span><span> </span><span>foo</span>

When considering whether a given char should be added onto the
end of an existing span or not, we were looking for the previous
non-whitespace char as the place to extend from.

Sadly, in the case where we've already encountered a leading space,
that meant we'd create another span needlessly.

Instead, modify the code to accept leading whitespace in spans.