Avoid mixing paragraphs.
This paragraph should       This paragraph should
not be mixed with the       not be mixed with the
one to the right when       one to the left when
we extract the file.        we extract the file.

During make_paragraphs, check for the horizontal
extend of paragraphs to ensure we don't end up
interleaving lines from different columns into
single paragraphs.